MXS-2185: Enable most -Wextra warnings
Enabled all -Wextra warnings except -Wunused-parameter and
-Wmissing-field-initializer.

The other "unused" parameter warnings are already disabled so the
-Wunused-parameter that -Wextra enables must be disabled.

The behavior that -Wmissing-field-initializer warns about is
well-defined behavior for both C and C++ and as such would require
writing constructors for all POD structs that are partially initialized,
even if zero-initialization would correctly initialize all class type
members.

+++ server/modules/parser_plugin/pp_sqlite/sqlite-src-3110100/src/expr.c
@@ -3009,7 +3009,7 @@ int sqlite3ExprCodeTarget(Parse *pParse, Expr *pExpr, int target){
         ** loading.
         */
         if( (pDef->funcFlags & (SQLITE_FUNC_LENGTH|SQLITE_FUNC_TYPEOF))!=0 ){
-          u8 exprOp;
+          u16 exprOp;
           assert( nFarg==1 );
           assert( pFarg->a[0].pExpr!=0 );
           exprOp = pFarg->a[0].pExpr->op;
